dev_requirements.txt
pillow
pandas
numpy
transformers
torch
tqdm
